:recycle: realocando pdf e pacienteslistar para hooks
